# arlo-admin-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  arlo-admin_internal_modules_file_dto.FileResponse:
    properties:
      category:
        description: '文件分类: image/video/audio/document/other'
        example: image
        type: string
      createdAt:
        description: 上传时间
        type: string
      id:
        description: 文件ID
        example: 1
        type: integer
      md5:
        description: MD5校验值
        example: d41d8cd98f00b204e9800998ecf8427e
        type: string
      mimeType:
        description: MIME类型
        example: image/png
        type: string
      name:
        description: 原始文件名
        example: avatar.png
        type: string
      size:
        description: 文件大小(字节)
        example: 102400
        type: integer
      uploader:
        description: 上传者
        example: admin
        type: string
      uploaderId:
        description: 上传者ID
        example: 1
        type: integer
      url:
        description: 访问URL
        example: http://localhost:8080/api/v1/file/1/download
        type: string
    type: object
...
  arlo-admin_pkg_response.Response:
    properties:
      code:
        description: 响应码（200=成功，>0=业务错误）
        example: 200
        type: integer
      data:
        description: 响应数据（具体类型取决于接口）
        type: object
      msg:
        description: 响应消息
        example: success
        type: string
    type: object
host: localhost:8090
info:
  contact:
    name: arlo-admin
  description: 管理后台 API 接口文档
  title: arlo-admin API
  version: 1.0.0
paths:
...
  /file/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 上传的文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/arlo-admin_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/arlo-admin_internal_modules_file_dto.FileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/arlo-admin_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 上传文件
      tags:
      - File
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"